:sparkles: Implement recursive apparent power calculation
